moved relevant methods to predict.py for modularity =
