+++ recognition/alzheimers_transformer_46963765/train.py
@@ -2,13 +2,14 @@ import dataset
 from modules import *
 import torch.optim as optim
 import matplotlib.pyplot as plt
-
+from predict import test_accuracy, visualize_loss
 
 
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
 #put in training loop
 def train(epochs, depth):
+    
     model = ADNI_Transformer(depth=depth)
     model.to(device=device)
 
@@ -16,9 +17,9 @@ def train(epochs, depth):
     train_loader = dataset.get_train_loader()
     optimizer = optim.Adam(model.parameters(), 0.002)
     criterion = nn.BCELoss()
-
     batch_losses = []
 
+    # training loop
     for epoch in range(epochs):
         batch_loss = 0
 
@@ -42,42 +43,6 @@ def train(epochs, depth):
     return model, batch_losses
 
 
-def visualize_loss(batch_losses):
-    
-    epochs = range(1, len(batch_losses) + 1)
-
-    
-    plt.plot(epochs, batch_losses, marker='o', linestyle='-')
-    plt.xlabel('Epoch')
-    plt.ylabel('Batch Loss')
-    plt.title('Batch Loss Over Epochs')
-    plt.grid(True)
-    plt.savefig('plots/loss_plot.png')
-
-    plt.show()
-
-
-def test_accuracy(model):
-    dataset = ds.ADNI_Dataset()
-    test_laoder = dataset.get_test_loader()
-        
-    correct_predictions = 0
-    total_samples = 0    
-        
-    model.eval() 
-    for j, (images, labels) in  enumerate(test_laoder):
-        if images.size(0) == 32:
-
-            images = images.to(device) 
-            labels = labels.to(device)
-            
-            outputs = model(images)
-            predictions = (outputs >= 0.5).squeeze().long()
-            correct_predictions += (predictions == labels).sum().item()
-            total_samples += labels.size(0)
-
-    accuracy = correct_predictions / total_samples
-    return accuracy
 
 if __name__ == "__main__":
     epochs = 75
